Preinstall fewer locales
The default from eos-image-builder is to install the following
locales for apps:

- ar (Arabic)
- be (Belarusian)
- bn (Bengali)
- en (English)
- es (Spanish)
- fr (French)
- hu (Hungarian)
- id (Indonesian)
- ko (Korean)
- pt (Portuguese)
- ro (Romanian)
- ru (Russian)
- sr (Serbian)
- uk (Ukrainian)
- zh (Chinese)

While it is harmless to have additional locales installed, it wastes
disk space, and bandwidth when installing updates.

This option only affects Flatpak apps and runtimes. All extant
translations for the desktop and built-in apps are always available; if
a user configures their account to use a language not on this list, the
app center will update apps to fetch those translations in due course.

(I have selected English and Spanish because according to
https://en.wikipedia.org/wiki/North_Carolina#Languages other languages
are spoken by less than 0.5% of the population of North Carolina.)