Correctly remove Chromium
In f00ec28d1bc9e5637e32f137508af0a753e68862 ("Move Kramden-specific
settings from defaults.ini") I incorrectly added the config to avoid
installing Chromium to the eos-apps section; but this app comes from
Flathub.

Move it to the flathub section, and also mark its .Codecs extension not
to be installed. (This extension is explicitly listed for installation
in the eos-image-builder config.)